+++ ErsatzTV.Application/Channels/Commands/CreateChannelHandler.cs
@@ -21,7 +21,7 @@ public class CreateChannelHandler : IRequestHandler<CreateChannel, Either<BaseEr
     {
         await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync(cancellationToken);
         Validation<BaseError, Channel> validation = await Validate(dbContext, request);
-        return await LanguageExtensions.Apply(validation, c => PersistChannel(dbContext, c));
+        return await validation.Apply(c => PersistChannel(dbContext, c));
     }
 
     private static async Task<CreateChannelResult> PersistChannel(TvContext dbContext, Channel channel)
@@ -36,6 +36,7 @@ public class CreateChannelHandler : IRequestHandler<CreateChannel, Either<BaseEr
             await FFmpegProfileMustExist(dbContext, request),
             ValidatePreferredAudioLanguage(request),
             ValidatePreferredSubtitleLanguage(request),
+            ValidateSubtitleAndMusicCredits(request),
             await WatermarkMustExist(dbContext, request),
             await FillerPresetMustExist(dbContext, request))
         .Apply(
@@ -45,6 +46,7 @@ public class CreateChannelHandler : IRequestHandler<CreateChannel, Either<BaseEr
                 ffmpegProfileId,
                 preferredAudioLanguageCode,
                 preferredSubtitleLanguageCode,
+                _,
                 watermarkId,
                 fillerPresetId) =>
             {
@@ -72,7 +74,8 @@ public class CreateChannelHandler : IRequestHandler<CreateChannel, Either<BaseEr
                     Artwork = artwork,
                     PreferredAudioLanguageCode = preferredAudioLanguageCode,
                     PreferredSubtitleLanguageCode = preferredSubtitleLanguageCode,
-                    SubtitleMode = request.SubtitleMode
+                    SubtitleMode = request.SubtitleMode,
+                    MusicVideoCreditsMode = request.MusicVideoCreditsMode
                 };
 
                 foreach (int id in watermarkId)
@@ -105,6 +108,17 @@ public class CreateChannelHandler : IRequestHandler<CreateChannel, Either<BaseEr
                 lc => string.IsNullOrWhiteSpace(lc) || CultureInfo.GetCultures(CultureTypes.NeutralCultures).Any(
                     ci => string.Equals(ci.ThreeLetterISOLanguageName, lc, StringComparison.OrdinalIgnoreCase)))
             .ToValidation<BaseError>("Preferred subtitle language code is invalid");
+
+    private static Validation<BaseError, string> ValidateSubtitleAndMusicCredits(CreateChannel createChannel)
+    {
+        if (createChannel.MusicVideoCreditsMode != ChannelMusicVideoCreditsMode.None &&
+            createChannel.SubtitleMode == ChannelSubtitleMode.None)
+        {
+            return BaseError.New("Subtitles are required for music video credits");
+        }
+
+        return string.Empty;
+    }
 
     private static async Task<Validation<BaseError, string>> ValidateNumber(
         TvContext dbContext,