+++ ErsatzTV.Core/FFmpeg/SubtitleBuilder.cs
@@ -9,6 +9,7 @@ public class SubtitleBuilder
     private Option<int> _alignment;
     private Option<int> _borderStyle;
     private string _content;
+    private Option<TimeSpan> _end;
     private Option<string> _fontName;
     private Option<int> _fontSize;
     private int _marginLeft;
@@ -18,6 +19,7 @@ public class SubtitleBuilder
     private Option<string> _primaryColor;
     private Option<IDisplaySize> _resolution = None;
     private Option<int> _shadow;
+    private Option<TimeSpan> _start;
 
     public SubtitleBuilder(ITempFilePool tempFilePool) => _tempFilePool = tempFilePool;
 
@@ -93,6 +95,13 @@ public class SubtitleBuilder
         return this;
     }
 
+    public SubtitleBuilder WithStartEnd(TimeSpan start, TimeSpan end)
+    {
+        _start = start;
+        _end = end;
+        return this;
+    }
+
     public async Task<string> BuildFile()
     {
         string fileName = _tempFilePool.GetNextTempFile(TempFileCategory.Subtitle);
@@ -116,15 +125,22 @@ public class SubtitleBuilder
         sb.AppendLine(
             $"Style: Default,{await _fontName.IfNoneAsync("")},{await _fontSize.IfNoneAsync(32)},{await _primaryColor.IfNoneAsync("")},{await _outlineColor.IfNoneAsync("")},{await _borderStyle.IfNoneAsync(0)},1,{await _shadow.IfNoneAsync(0)},{await _alignment.IfNoneAsync(0)},1");
 
+        var start = "0:00:00.00";
+        foreach (TimeSpan startTime in _start)
+        {
+            start = $"{(int)startTime.TotalHours:00}:{startTime.ToString(@"mm\:ss\.ff")}";
+        }
+
+        var end = "99:99:99.99";
+        foreach (TimeSpan endTime in _end)
+        {
+            end = $"{(int)endTime.TotalHours:00}:{endTime.ToString(@"mm\:ss\.ff")}";
+        }
+
         sb.AppendLine("[Events]");
         sb.AppendLine("Format: Layer, Start, End, Style, Name, MarginL, MarginR, MarginV, Effect, Text");
         sb.AppendLine(
-            $"Dialogue: 0,0:00:00.00,99:99:99.99,Default,,{_marginLeft},{_marginRight},{_marginV},,{_content}");
-
-        if (!string.IsNullOrWhiteSpace(_content))
-        {
-            sb.AppendLine(_content);
-        }
+            @$"Dialogue: 0,{start},{end},Default,,{_marginLeft},{_marginRight},{_marginV},,{{\fad(1200,1200)}}{_content}");
 
         await File.WriteAllTextAsync(fileName, sb.ToString());